Rename .util.logging to ._logging to avoid shadowing standard library
