+++ message_ix_models/__init__.py
@@ -4,7 +4,7 @@ from pkg_resources import DistributionNotFound, get_distribution
 
 from message_ix_models.util.context import Context
 from message_ix_models.util.importlib import MessageDataFinder
-from message_ix_models.util.logging import setup as setup_logging
+from message_ix_models.util._logging import setup as setup_logging
 from message_ix_models.util.scenarioinfo import ScenarioInfo
 
 # Expose utility classes
