+++ message_ix_models/cli.py
@@ -21,8 +21,8 @@ import click
 
 from message_ix_models.util.click import common_params
 from message_ix_models.util.context import Context
-from message_ix_models.util.logging import mark_time
-from message_ix_models.util.logging import setup as setup_logging
+from message_ix_models.util._logging import mark_time
+from message_ix_models.util._logging import setup as setup_logging
 
 log = logging.getLogger(__name__)
