+++ message_ix_models/tests/util/test_logging.py
@@ -1,7 +1,7 @@
 import logging
 import re
 
-from message_ix_models.util.logging import mark_time, silence_log
+from message_ix_models.util._logging import mark_time, silence_log
 
 
 def test_mark_time(caplog):
